requirements.txt
fastapi 
uvicorn 
requests 
python-dotenv
uvicorn
python-dotenv
langchain
langchain_groq
pydantic
openai
